+++ src/routes/BaseAliasRoute.ts
@@ -9,6 +9,7 @@ import {
     DocDecorators,
     HttpRequest,
     ObjectFactory,
+    RateLimiter,
     RepoUtils,
     RouteDecorators,
     UpdateObject,
@@ -16,7 +17,6 @@ import {
 import { ApiError, JWTUser, MessagingUtils, ObjectDecorators, UserUtils, ValidationUtils } from "@rapidrest/core";
 import { Alias, AliasType, ContactType, Profile } from "../models/types.js";
 import { generateOTP, verifyOTP } from "../auth/shared.js";
-import { RateLimiter } from "../auth/RateLimiter.js";
 
 const { Config, Init, Inject } = ObjectDecorators;
 const { Description, Returns, Summary } = DocDecorators;