+++ src/routes/BaseOAuthRevokeRoute.ts
@@ -4,14 +4,13 @@
 ///////////////////////////////////////////////////////////////////////////////
 import "reflect-metadata";
 import { ObjectDecorators } from "@rapidrest/core";
-import { DocDecorators, HttpRequest, HttpResponse, ObjectFactory, RepoUtils, RouteDecorators } from "@rapidrest/service-core";
+import { DocDecorators, HttpRequest, HttpResponse, ObjectFactory, RateLimiter, RepoUtils, RouteDecorators } from "@rapidrest/service-core";
 import { Client, OAuthRefreshToken, SigningKey } from "../models/types.js";
 import { AccessTokenDenylist } from "../auth/AccessTokenDenylist.js";
 import { ClientAuthUtils } from "../auth/ClientAuthUtils.js";
 import { OAuthError, toOAuthError } from "../auth/OAuthError.js";
 import { OAuthTokenUtils } from "../auth/OAuthTokenUtils.js";
 import { SigningKeyUtils } from "../auth/SigningKeyUtils.js";
-import { RateLimiter } from "../auth/RateLimiter.js";
 import { getRequestData, hashOpaqueToken } from "../auth/shared.js";
 
 const { Init, Inject } = ObjectDecorators;